Update view spacing/fonts, update code formatting, add some new logic to match ios implementation

+++ app/src/main/java/com/example/healthAppAndroid/homeTab/addWorkout/MutableString.java
@@ -2,8 +2,8 @@ package com.example.healthAppAndroid.homeTab.addWorkout;
 
 final class MutableString {
     @SuppressWarnings("StringBufferField") final StringBuilder str = new StringBuilder(16);
-    short index = 0;
-    short end = 1;
+    int index = 0;
+    int end = 1;
 
     void replace(String replacement) { str.replace(index, end, replacement); }
 }
